Two fixes to Prince's behaviour:
Previously it could consider itself as a valid target to play both when determining priority and when playing if there were cost reducers in play.
Previously it would ignore the first card in the hand when trying to determine whether it wanted to be played.

Added test for the second of these.